+++ src/streetworks/common/__init__.py
@@ -39,6 +39,7 @@ from .from_lisboa import from_lisboa
 from .from_madrid import from_madrid
 from .from_mallorca import from_mallorca
 from .from_milano import from_milano
+from .from_monaghan import from_monaghan_road
 from .from_nrn import from_nrn
 from .from_nsw_livetraffic import from_nsw_livetraffic
 from .from_nvdb import from_nvdb
@@ -164,4 +165,5 @@ __all__ = [
     "from_guernsey_street",
     "from_nrn",
     "from_gibraltar_street",
+    "from_monaghan_road",
 ]